requirements.txt
absl-py==0.15.0
albumentations==1.3.0
altgraph==0.17.4
appdirs==1.4.4
astunparse==1.6.3
backcall ==0.2.0
backports.functools-lru-cache==1.6.5
cached-property==1.5.2
cachetools==5.0.0
certifi==2023.5.7
charset-normalizer==2.0.12
click==8.1.7
colorama ==0.4.5
colorlog==6.7.0
colormap==1.0.4
convcolors==2.2.0
cycler==0.11.0
debugpy ==1.6.3
decorator ==4.4.2
docker-pycreds==0.4.0
dominate==2.9.0
easydev==0.12.0
easydict==1.10
efficientnet-pytorch==0.7.1
entrypoints ==0.4
extcolors==1.0.0
filelock==3.12.0
flatbuffers==1.12
fonttools==4.32.0
fsspec==2023.1.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.6.5
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
huggingface-hub==0.15.1
idna==3.3
imagecodecs==2021.11.20
imagecodecs-lite==2022.9.26
imageio ==2.9.0
imageio-ffmpeg==0.4.8
imglyb==2.1.0
importlib-metadata==4.11.3
ipykernel ==6.16.2
ipython ==7.33.0
jedi ==0.18.2
jgo==1.0.5
joblib ==1.1.1
JPype1==1.4.1
jsonpatch==1.33
jsonpointer==2.4
jupyter_client ==7.4.9
jupyter_core ==4.11.1
keras-nightly==2.5.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
labeling==0.1.13
libtiff ==4.2.0
Markdown==3.3.6
matplotlib==3.0.3
matplotlib-inline ==0.1.6
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
moviepy==1.0.3
munch==3.0.0
nest-asyncio ==1.5.6
networkx==2.6.3
numpy==1.19.5
oauthlib==3.2.0
olefile==0.46
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
packaging ==21.3
pandas ==1.2.4
parso ==0.8.3
pexpect==4.8.0
pickleshare ==0.7.5
Pillow==9.1.0
pretrainedmodels==0.7.4
proglog==0.1.10
prompt-toolkit ==3.0.39
protobuf==3.20.0
psutil ==5.9.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments ==2.15.1
pyimagej==1.0.0
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2024.0
pyjnius==1.6.1
pyparsing==3.0.8
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil ==2.8.2
PyWavelets==1.3.0
pywin32==303
pywin32-ctypes==0.2.2
PyYAML==6.0
pyzmq ==23.2.0
qudida==0.0.4
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
safetensors==0.3.1
scikit-image==0.19.3
scikit-learn ==1.0.2
scipy==1.7.3
scyjava==1.9.1
seaborn==0.11.0
segmentation-models-pytorch==0.3.3
sentry-sdk==1.38.0
setproctitle==1.3.3
six ==1.15.0
smmap==5.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
termcolor==1.1.0
thop==0.1.1
threadpoolctl ==2.2.0
tifffile==2021.11.2
timm==0.9.2
torch ==1.8.1
torchstat==0.0.7
torchvision ==0.9.1
tornado ==6.2
tqdm==4.65.0
traitlets ==5.9.0
typing_extensions==4.6.3
urllib3==1.26.18
visdom==0.2.4
wandb==0.16.1
wcwidth ==0.2.6
websocket-client==1.6.1
Werkzeug==2.1.1
wincertstore==0.2
wrapt==1.12.1
xarray==0.20.2
zipp==3.8.0
scikit-image==0.20.0
scikit-learn==1.1.1
xgboost==2.0.3
lightgbm==3.2.1


